% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big5}
\alias{big5}
\title{Personality questionary data}
\format{
A tibble with 19719 rows and 57 variables:
\describe{
\item{engat}{Response to "is English your native language?". 1=yes, 2=no (0=missed)}
\item{gender}{gender	Chosen from a drop down menu. 1=Male, 2=Female, 3=Other (0=missed)}
\item{hand}{"What hand do you use to write with?". 1=Right, 2=Left, 3=Both (0=missed)}
\item{country}{The participant's technical location. ISO country code.}
\item{source}{How the participant came to the test. Based on HTTP Referer.
1=from another page on the test website, 2=from google, 3=from facebook,
4=from any url with ".edu" in its domain name (e.g. xxx.edu, xxx.edu.au),
6=other source, or HTTP Referer not provided.}
\item{...}{+ the 50 questions rated on a five point scale where 1=Disagree,
3=Neutral, 5=Agree (0=missed)}
}
}
\source{
\url{https://openpsychometrics.org/_rawdata/}
}
\usage{
big5
}
\description{
Answers to 50 lickert rated statements from 19719 participants.
The test can be made here: \url{https://openpsychometrics.org/tests/IPIP-BFFM/}
}
\details{
This data was collected (c. 2012) through on interactive online personality
test. Participants were informed that their responses would be recorded and
used for research at the begining of the test and asked to confirm their
consent at the end of the test.
}
\keyword{datasets}
