% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplotly.R
\name{triplotly}
\alias{triplotly}
\title{PCA triplots}
\usage{
triplotly(
  data,
  factors,
  group_by,
  components = c(1, 2),
  alpha = 0,
  title = "",
  arr.scale = 1,
  scale.pc = F,
  colorPalette = "RdYlBu",
  opacity = 1
)
}
\arguments{
\item{data}{}

\item{factors}{}

\item{group_by}{}

\item{components}{}

\item{alpha}{}

\item{title}{}

\item{arr.scale}{}

\item{scale.pc}{}

\item{colorPalette}{}

\item{opacity}{}
}
\value{
a graphic
}
\description{
PCA triplots
}
