% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplotly.R
\name{trp}
\alias{trp}
\title{Triplotly constructor}
\description{
Triplotly constructor

Triplotly constructor
}
\examples{

## ------------------------------------------------
## Method `trp$new`
## ------------------------------------------------

library(triplotly)
data(iris)
pca1 <- trp$new(iris, factor = "Species")

Title


## ------------------------------------------------
## Method `trp$doSVD`
## ------------------------------------------------

Title


## ------------------------------------------------
## Method `trp$calcBi_df`
## ------------------------------------------------

Title

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-trp-new}{\code{trp$new()}}
\item \href{#method-trp-doSVD}{\code{trp$doSVD()}}
\item \href{#method-trp-calcBi_df}{\code{trp$calcBi_df()}}
\item \href{#method-trp-plot}{\code{trp$plot()}}
\item \href{#method-trp-clone}{\code{trp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trp-new"></a>}}
\if{latex}{\out{\hypertarget{method-trp-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trp$new(data, factors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe or tibble to perform PCA}

\item{\code{factors}}{The columns of the dataframe that should be excluded from PCA,
either as integer or string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class "trp" with different functions to explore the data
via PCA tri- or biplots.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(triplotly)
data(iris)
pca1 <- trp$new(iris, factor = "Species")

Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trp-doSVD"></a>}}
\if{latex}{\out{\hypertarget{method-trp-doSVD}{}}}
\subsection{Method \code{doSVD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trp$doSVD()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trp-calcBi_df"></a>}}
\if{latex}{\out{\hypertarget{method-trp-calcBi_df}{}}}
\subsection{Method \code{calcBi_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trp$calcBi_df(components, group_by, alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{}

\item{\code{group_by}}{}

\item{\code{alpha}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trp-plot"></a>}}
\if{latex}{\out{\hypertarget{method-trp-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trp$plot(
  arr.scale = 1,
  scale.pc = FALSE,
  colorPalette = "RdYlBu",
  opacity = 1,
  size = 1,
  showLabels = FALSE,
  title = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arr.scale}}{}

\item{\code{scale.pc}}{}

\item{\code{colorPalette}}{}

\item{\code{opacity}}{}

\item{\code{size}}{}

\item{\code{showLabels}}{}

\item{\code{title}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-trp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
